Front end and back end - big changes -  fixing the "DST fudging", because it was fatally flawed since I introduced it in revision fa5994f77bb7 (3 weeks ago).
The flaw was that we were fudging the forecast target times, but not the
observation times.  So for example for a WindFinder Regular forecast in the
winter for 7:00 AM, we were comparing that to the actual wind at 8:00 AM.

Also, the new fudging system is done after selecting forecasts from the
database.  (The old one was done before inserting forecasts into the database.)

Also, this revision also removes the PNG graph from the main web page.  The
informoation is all in text now.