Main web page - changing caching, because of a bug that I think I saw but couldn't reproduce, regarding caching of our AJAX call to get the .json file which contains the stats in HTML.
I saw (or think I saw) this bug on the live site I saw some table (in the HTML
table) which was a few days old.  Refreshing the page in the browser wouldn't
update it, but Chrome's "empty cache and hard reload" did.

Even if what I saw wasn't a bug, or it was a bug and this revision doesn't fix
- I hope at least that this revision doesn't do any harm.